+++ src/features/Character/components/Skill/index.js
@@ -7,33 +7,39 @@ import T from 'i18n-react';
 
 import TooltipTrigger from 'common/components/TooltipTrigger';
 import Icon from 'common/components/Icon';
+import ResourceLink, { buildLink } from 'common/components/ResourceLink';
 
 import EmptySkill from './Empty';
 import styles from './styles.less';
 
 type Props = {
-  data?: {},
+  data?: {
+    name: string,
+  },
   className?: string,
   tooltipTextOverride?: string,
   size?: number,
+  inlineText?: string,
 };
 
-const Skill = ({ data, className, tooltipTextOverride, size }: Props) => {
+const Skill = ({ data, className, tooltipTextOverride, size, inlineText }: Props) => {
   const error = get(data, 'error');
   const tooltipData = tooltipTextOverride || data || T.translate('characters.noSkill');
 
   return (
     <TooltipTrigger type="skill" data={tooltipData}>
-      {(error || !data)
-        ? <EmptySkill size={size} />
-        : (
-          <Icon
-            src={get(data, 'icon')}
-            size="mediumSmall"
-            className={cx(styles.skill, className)}
-            sizePx={size}
-          />
-        )}
+      <ResourceLink text={data && data.name} href={buildLink(inlineText, data && data.name)}>
+        {(error || !data)
+          ? <EmptySkill size={size} />
+          : (
+            <Icon
+              src={get(data, 'icon')}
+              size="mediumSmall"
+              className={cx(styles.skill, className)}
+              sizePx={size}
+            />
+          )}
+      </ResourceLink>
     </TooltipTrigger>
   );
 };