+++ src/features/Character/components/Trait/index.spec.js
@@ -19,6 +19,10 @@ const Trait = proxyquire('features/Character/components/Trait', {
   'common/components/TooltipTrigger': TooltipTrigger,
   'common/components/Icon': Icon,
   'common/styles/colours': colours,
+  'common/components/ResourceLink': {
+    default: (props) => <div {...props} />,
+    buildLink: () => '',
+  },
 });
 
 describe('<Trait />', () => {